+++ src/main/java/com/springkafka/kafka_app/utils/serdes/HashMapSerializerDeserializer.java
@@ -8,14 +8,14 @@ import org.apache.kafka.common.serialization.Serializer;
 import java.util.HashMap;
 import java.util.Map;
 
-public class HashMapSerializerDeserializer implements Serializer<Map<String, Integer>>, Deserializer<Map<String, Integer>> {
+public class HashMapSerializerDeserializer implements Serializer<Map<String, Long>>, Deserializer<Map<String, Long>> {
 
     @Override
     public void configure(Map<String, ?> configs, boolean isKey) {
     }
 
     @Override
-    public byte[] serialize(String topic, Map<String, Integer> data) {
+    public byte[] serialize(String topic, Map<String, Long> data) {
 
         try {
             return ServiceProperties.objectmapper.writeValueAsBytes(data);
@@ -26,11 +26,10 @@ public class HashMapSerializerDeserializer implements Serializer<Map<String, Int
     }
 
     @Override
-    public HashMap<String, Integer> deserialize(String topic, byte[] data) {
+    public HashMap<String, Long> deserialize(String topic, byte[] data) {
         try {
-            return (HashMap<String, Integer>)ServiceProperties.objectmapper.readValue(data,HashMap.class);
-//            HashMap<String, Integer> hashMap = ServiceProperties.objectmapper.readValue(data,HashMap.class);
-//            return hashMap;
+            return (HashMap<String, Long>)ServiceProperties.objectmapper.readValue(data,HashMap.class);
+
         } catch (Exception e) {
             e.printStackTrace();
             return null;
